@@ -1,11 +1,11 @@
 name: route_handler
 description: A simple abstraction that provides a structured guideline for named route handling in Flutter
-version: 1.0.2
+version: 2.0.0
 homepage: https://github.com/magicleon94/route_handler
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.17.0"
 
 dependencies:
   flutter:
